tienda/requirements.txt
Django==5.0.1
mercadopago==2.2.0
python-dotenv==1.0.0
mysql-connector-python==8.3.0